refactor: added markdown and regression module
